requirement.txt
opencv-python
Pillow
pyocr
selenium